final fix for captioning
